# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "BibleSearchResponse": {
        "additionalProperties": true,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/BibleVerse"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "query",
          "count",
          "results"
        ],
        "title": "BibleSearchResponse",
        "type": "object"
      },
      "BibleVerse": {
        "additionalProperties": true,
        "properties": {
          "book_english": {
            "title": "Book English",
            "type": "string"
          },
          "book_id": {
            "title": "Book Id",
            "type": "integer"
          },
          "book_zomi": {
            "title": "Book Zomi",
            "type": "string"
          },
          "changes_applied": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changes Applied"
          },
          "chapter": {
            "title": "Chapter",
            "type": "integer"
          },
          "english": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "English"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "testament": {
            "title": "Testament",
            "type": "string"
          },
          "verse": {
            "title": "Verse",
            "type": "integer"
          },
          "zomi_iso": {
            "title": "Zomi Iso",
            "type": "string"
          },
          "zomi_original": {
            "title": "Zomi Original",
            "type": "string"
          }
        },
        "required": [
          "testament",
          "book_id",
          "book_english",
          "book_zomi",
          "chapter",
          "verse",
          "reference",
          "zomi_original",
          "zomi_iso"
        ],
        "title": "BibleVerse",
        "type": "object"
      },
      "DictionaryEntry": {
        "additionalProperties": true,
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "definition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition"
          },
          "dialect": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dialect"
          },
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "english": {
            "title": "English",
            "type": "string"
          },
          "examples": {
            "items": {
              "$ref": "#/components/schemas/TranslateMatch"
            },
            "title": "Examples",
            "type": "array"
          },
          "flagged": {
            "default": false,
            "title": "Flagged",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "letter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Letter"
          },
          "normalized_word": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Word"
          },
          "part_of_speech": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Of Speech"
          },
          "part_of_speech_full": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Of Speech Full"
          },
          "pronunciation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pronunciation"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "source_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference"
          },
          "source_word": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Word"
          },
          "suggested_correction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Correction"
          },
          "usage_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usage Notes"
          },
          "verification_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Status"
          },
          "verified": {
            "default": false,
            "title": "Verified",
            "type": "boolean"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "zomi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zomi"
          }
        },
        "required": [
          "id",
          "english"
        ],
        "title": "DictionaryEntry",
        "type": "object"
      },
...
      "DictionarySearchResponse": {
        "additionalProperties": true,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "direction": {
            "default": "en-zom",
            "title": "Direction",
            "type": "string"
          },
          "letter": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Letter"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/DictionaryEntry"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "query",
          "count",
          "results"
        ],
        "title": "DictionarySearchResponse",
        "type": "object"
      },
      "DictionarySuggestion": {
        "additionalProperties": true,
        "properties": {
          "definition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Definition"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "part_of_speech": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Part Of Speech"
          },
          "source_word": {
            "title": "Source Word",
            "type": "string"
          },
          "verification_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verification Status"
          }
        },
        "required": [
          "id",
          "source_word"
        ],
        "title": "DictionarySuggestion",
        "type": "object"
      },
      "DictionarySuggestionsResponse": {
        "additionalProperties": true,
        "properties": {
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/DictionarySuggestion"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "query",
          "results"
        ],
        "title": "DictionarySuggestionsResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TranslationSearchResponse": {
        "additionalProperties": true,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/TranslateMatch"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "query",
          "count",
          "results"
        ],
        "title": "TranslationSearchResponse",
        "type": "object"
      },
      "UnifiedSearchResponse": {
        "additionalProperties": true,
        "properties": {
          "bible": {
            "items": {
              "$ref": "#/components/schemas/BibleVerse"
            },
            "title": "Bible",
            "type": "array"
          },
          "dictionary": {
            "items": {
              "$ref": "#/components/schemas/DictionaryEntry"
            },
            "title": "Dictionary",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "translations": {
            "items": {
              "$ref": "#/components/schemas/TranslateMatch"
            },
            "title": "Translations",
            "type": "array"
          }
        },
        "required": [
          "query",
          "dictionary",
          "translations",
          "bible"
        ],
        "title": "UnifiedSearchResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Zomi language platform — dictionary, translation retrieval, Bible, and learning.",
    "title": "Siamsil API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/bible/search": {
      "get": {
        "operationId": "search_bible_api_bible_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BibleSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Bible",
        "tags": [
          "bible"
        ]
      }
    },
    "/api/dictionary": {
      "get": {
        "operationId": "search_dictionary_api_dictionary_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "letter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Letter"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "en-zom",
              "pattern": "^(en-zom|zom-en)$",
              "title": "Direction",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictionarySearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Dictionary",
        "tags": [
          "dictionary"
        ]
      }
    },
...
    "/api/dictionary/suggest": {
      "get": {
        "operationId": "suggest_api_dictionary_suggest_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictionarySuggestionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Suggest",
        "tags": [
          "dictionary"
        ]
      }
    },
...
    "/api/dictionary/{entry_id}": {
      "get": {
        "operationId": "get_entry_api_dictionary__entry_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictionaryEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Entry",
        "tags": [
          "dictionary"
        ]
      }
    },
...
    "/api/search": {
      "get": {
        "operationId": "unified_search_api_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unified Search",
        "tags": [
          "search"
        ]
      }
    },
...
    "/api/translate/search": {
      "get": {
        "operationId": "search_matches_api_translate_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 500,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Matches",
        "tags": [
          "translate"
        ]
      }
    },
...
    "/api/v1/bible/search": {
      "get": {
        "operationId": "search_bible_api_v1_bible_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BibleSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Bible",
        "tags": [
          "bible"
        ]
      }
    },
    "/api/v1/dictionary": {
      "get": {
        "operationId": "search_dictionary_api_v1_dictionary_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "letter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Letter"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "en-zom",
              "pattern": "^(en-zom|zom-en)$",
              "title": "Direction",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictionarySearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Dictionary",
        "tags": [
          "dictionary"
        ]
      }
    },
...
    "/api/v1/dictionary/suggest": {
      "get": {
        "operationId": "suggest_api_v1_dictionary_suggest_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 80,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictionarySuggestionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Suggest",
        "tags": [
          "dictionary"
        ]
      }
    },
...
    "/api/v1/dictionary/{entry_id}": {
      "get": {
        "operationId": "get_entry_api_v1_dictionary__entry_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictionaryEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Entry",
        "tags": [
          "dictionary"
        ]
      }
    },
...
    "/api/v1/search": {
      "get": {
        "operationId": "unified_search_api_v1_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 8,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unified Search",
        "tags": [
          "search"
        ]
      }
    },
...
    "/api/v1/translate/search": {
      "get": {
        "operationId": "search_matches_api_v1_translate_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 500,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Matches",
        "tags": [
          "translate"
        ]
      }
    },
...
  }
}